Test: Split tests
